Refactor zAppBuild reporting capabilities (#341)
Moved the Reporting feature into in its own utilities file, and simplifying the query to check for impacted logical files across collections using DBB 2.0 APIs.
